+++ bindings/python/iota_sdk/__init__.py
@@ -19,6 +19,7 @@ from .types.common import *
 from .types.event import *
 from .types.feature import *
 from .types.filter_options import *
+from .types.input import *
 from .types.native_token import *
 from .types.network_info import *
 from .types.node_info import *
@@ -30,5 +31,7 @@ from .types.payload import *
 from .types.send_params import *
 from .types.token_scheme import *
 from .types.transaction import *
+from .types.transaction_data import *
 from .types.transaction_options import *
+from .types.unlock import *
 from .types.unlock_condition import *